🎨 Factorization
